Fix subtle issue -- parser needs end signal
To distinguish between "123(EOF)" and
"123..", which could also become "1234".
Note that this would be a problem over a
stream socket, so it is NOT safe to send
raw sexps over a stream and wait for a
response. Appending an extra space would
technically help, but feels like an ugly
hack. The proper "hack" is to simply
send lists only.

(The plan is that each list begins with
a symbol which specifies which handler
the cdr goes to.)