Avoid messing with srand(), use MT
MT code not by me, freeware, license in code.
Note that generating the random number each frame
without any sort of caching is not a sane way of
doing things, and it starts to show at this point
since MT has a much more involved initialization
phase than the usual implementations of standard
srand(). This was more or less just a
theoretically-cute way to generate an infinite
world for testing. (I'm surprised it looks
this good!)