New binary to gitignore
Also, note to self: drawing tile by tile is the
bottleneck right now, and it's pretty slow,
though at this stage that isn't a priority.
Later, to get CPU usage down, putting the
tiles in a display list and putting the level
sprites on one or a few sprite sheets is the
key.